backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.4
